+++ test/runtests.jl
@@ -2,6 +2,7 @@ using CliffordNumbers
 using Aqua, Test
 using LinearAlgebra, Quaternions, StaticArrays, Unitful
 using Adapt
+using Random
 
 Aqua.test_all(CliffordNumbers; unbound_args = false)
 
@@ -21,6 +22,7 @@ end
     include("indexing.jl")
     include("conversion.jl")
     include("operations.jl")
+    include("custom_signature.jl")
     include("ext/Adapt.jl")
     include("ext/LinearAlgebra.jl")
     include("ext/Quaternions.jl")